Move massage_sessions_result tests to tests/snuba/test_outcomes.py
The massage tests belong alongside the code they test, which now lives
in snuba/outcomes.py. This leaves test_sessions_v2.py focused on
QueryDefinition, get_constrained_date_range, and get_timestamps.
